Move Swipe Callback to an abstract class
